+++ transports/webrtc-websys/src/upgrade.rs
@@ -1,4 +1,4 @@
-use std::net::SocketAddr;
+use std::{net::SocketAddr, num::NonZeroUsize};
 
 use libp2p_identity::{Keypair, PeerId};
 use libp2p_webrtc_utils::{Fingerprint, StreamConfig, noise};
@@ -14,12 +14,14 @@ pub(crate) async fn outbound(
     remote_fingerprint: Fingerprint,
     id_keys: Keypair,
     stream_config: StreamConfig,
+    max_read_buffer_size: NonZeroUsize,
 ) -> Result<(PeerId, Connection), Error> {
     let fut = SendWrapper::new(outbound_inner(
         sock_addr,
         remote_fingerprint,
         id_keys,
         stream_config,
+        max_read_buffer_size,
     ));
     fut.await
 }
@@ -30,12 +32,14 @@ async fn outbound_inner(
     remote_fingerprint: Fingerprint,
     id_keys: Keypair,
     stream_config: StreamConfig,
+    max_read_buffer_size: NonZeroUsize,
 ) -> Result<(PeerId, Connection), Error> {
     let rtc_peer_connection = RtcPeerConnection::new(remote_fingerprint.algorithm()).await?;
 
     // Create stream for Noise handshake
     // Must create data channel before Offer is created for it to be included in the SDP
-    let (channel, listener) = rtc_peer_connection.new_handshake_stream(stream_config);
+    let (channel, listener) =
+        rtc_peer_connection.new_handshake_stream(stream_config, max_read_buffer_size);
     drop(listener);
 
     let ufrag = libp2p_webrtc_utils::sdp::random_ufrag();
@@ -66,5 +70,8 @@ async fn outbound_inner(
 
     tracing::debug!(peer=%peer_id, "Remote peer identified");
 
-    Ok((peer_id, Connection::new(rtc_peer_connection, stream_config)))
+    Ok((
+        peer_id,
+        Connection::new(rtc_peer_connection, stream_config, max_read_buffer_size),
+    ))
 }